Changed the contentMode of the image in the Details Screen to scaleAspectFill.
NOTE: This commit was done after verifying that this change had no impact on the scrolling or the functionality of the app.
While testing, I noticed that the "Close" button in the Image Details screen was sometimes coming in blank space outside the loaded image. Changing the contentMode of the image in the Details Screen to scaleAspectFill fixes this issue.